renaming actions to action
